### 2024-05-30 #6809 @editorialbot
[PRE REVIEW]: solartoolbox: A Python package for spatial solar1 energy analyses
**Submitting author:** @jranalli (<a href="http://orcid.org/0000-0002-8184-9895">Joseph Ranalli</a>)
**Repository:** https://github.com/jranalli/solartoolbox
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.4.0
