### 2024-05-30 #6809 @kandersolar
According to JOSS's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy), I have a COI with one of the authors: Will Hobbs and I have coauthored several publications within the last four years, and I expect we will continue to publish together in the future.  I believe I can provide an impartial review despite the COI, so I will be happy to review this submission if JOSS is willing to waive the conflict.

FYI I expect I will not have availability to complete a review for at least a month or perhaps two.
