### 2024-07-09 #6809 @EwaGomez
@AdamRJensen Unfortunatly I don't have the capacity for a JOSS review right now. Sorry! If you're still searching for reviewers mid of august, I would be happy to review!
