### 2024-07-09 #6809 @cwhanse
> @cwhanse it'd be great if you could serve as a reviewer. Thanks for informing us of the COI, you can consider it waived.

Certainly. We are only looking at an advance copy of the paper at this stage? If so, then I have read the draft paper and have no comments.
