### 2024-07-10 #6809 @AdamRJensen
> Certainly. We are only looking at an advance copy of the paper at this stage? If so, then I have read the draft paper and have no comments.
@cwhanse as soon as we secure a second reviewer I'll go ahead and start the review process. In short, you'll have a checklist with specific actions to go through, that includes checking the paper (which you've already done).
