### 2024-07-10 #6809 @AdamRJensen
@jranalli could you suggest some more reviewers? Perhaps some outside of the JOSS reviewers list, e.g., active people within the solar energy FOSS world.
