### 2024-07-10 #6809 @jranalli
I don't have a huge network of solar FOSS folks, but searching through some package lists and contribs to packages I know of, possibly the following:
- kperrynrel
- mhdella
- echedey-ls
- kbonney
- sjpfenninger
- LukasBommes

There are two additional suggestions I know of from DLR, but I wasn't able to find them as users on GitHub. Could provide emails if that's the route you'd want to go.
