### 2024-07-12 #6809 @AdamRJensen
> @AdamRJensen if you need me to act as a reviewer here, I can. Happy to sit this one out though if you have enough people

It'd be awesome if you could review 🎉  - I think it's right up your alley with your work on pvanalytics. Ideally, we have three reviewers in case one drops out
