### 2024-07-15 jranalli/solarspatialtools#7 @cwhanse
Documentation shortcomings
Below is the section about software documentation from the reviewer checklist for JOSS submissions. The only documentation is the README page.

I am hesitant to check the unchecked items. My concerns are pasted below each check box item.

### Documentation

The first paragraph of the README page should specify the scope of this package, which is tools for spatial analysis of solar-related data. This is clear from the paper.

Twelve "demos" are included; some are jupyter notebooks, others are python scripts. The notebooks are adequately commented. `cmv_demo.ipynb` does not run (won't preview on github). `cmv_demo.py` has minimal commenting.

It would be very helpful if the demos were rendered as documentation pages. Without any sort of introduction for new users, it is difficult to know where to enter these examples to learn how to use the tools in this package.

I don't think so. The primary failiing is the lack of a catalog of public functions. One must browse each code modules to discover what's available. Each function appears to be adequately documented, but checking again requires browsing each code file.

Without the benefit of API documentation, test coverage of public functions is difficult to judge.

More could be provided on this topic. The single sentence doesn't convey a commitment by maintainers to support the package into the future.
