### 2024-07-15 jranalli/solarspatialtools#8 @cwhanse
Package name
`solartoolbox` is quite generic - perhaps something like `solarspatialtools` would better convey the intended scope?
