### 2024-07-15 4ee010e
Very preliminary effort at starting sphinx docs
