+++ docs/sphinx/source/index.rst
@@ -1,0 +1,16 @@
+.. solartoolbox documentation master file, created by
+   sphinx-quickstart on Mon Jul 15 15:04:14 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+solartoolbox documentation
+==========================
+
+Add your content using ``reStructuredText`` syntax. See the
+`reStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_
+documentation for details.
+
+
+.. toctree::
+   api
+
