+++ docs/sphinx/source/cmv.rst
@@ -1,0 +1,7 @@
+solartoolbox.cmv module
+-----------------------
+
+.. automodule:: solartoolbox.cmv
+   :members:
+   :undoc-members:
+   :show-inheritance:
