+++ docs/sphinx/source/api.rst
@@ -1,0 +1,11 @@
+API
+===
+
+.. autosummary::
+    :toctree: generated
+
+    solartoolbox.cmv
+    solartoolbox.field
+
+.. toctree::
+    solartoolbox
