+++ docs/sphinx/source/solartoolbox.rst
@@ -1,0 +1,51 @@
+solartoolbox package
+====================
+
+.. CMV
+ ---
+ See the :py:mod:`~solartoolbox.cmv` module
+.. Field
+-----
+..
+See the :py:mod:`~solartoolbox.field` module
+
+
+solartoolbox.irradiance module
+------------------------------
+
+.. automodule:: solartoolbox.irradiance
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+solartoolbox.signalproc module
+------------------------------
+
+.. automodule:: solartoolbox.signalproc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+solartoolbox.spatial module
+---------------------------
+
+.. automodule:: solartoolbox.spatial
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+solartoolbox.stats module
+-------------------------
+
+.. automodule:: solartoolbox.stats
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: solartoolbox
+   :members:
+   :undoc-members:
+   :show-inheritance:
