+++ docs/sphinx/source/field.rst
@@ -1,0 +1,7 @@
+solartoolbox.field module
+-------------------------
+
+.. automodule:: solartoolbox.field
+   :members:
+   :undoc-members:
+   :show-inheritance:
