### 2024-07-16 945ffa9
Hopefully fix the github rendering of cmv_demo.ipynb, add comments to cmv_demo.py, in response to Issue #7
