+++ demos/cmv_demo.py
@@ -4,13 +4,33 @@ import matplotlib.pyplot as plt
 
 from solartoolbox import cmv, spatial
 
+# #############
+# # READ DATA #
+# #############
+
+# This is the name of the default datafile
+# It contains the definition of a plant layout as well as two time periods
+# of data with expected CMVs. These data consist of normalized combiner
+# current time series from a utility scale PV plant.
 datafile = "data/sample_plant_2.h5"
 pos_utm = pd.read_hdf(datafile, mode="r", key="utm")
 df = pd.read_hdf(datafile, mode="r", key="data_a")
 
+# The CMV routines in principle would work on non-normalized data, but have
+# some quality control parameters that are scaled to work with clearsky index.
+# So we'll normalize the current data to resemble clearsky index here. The
+# alternative would be to adjust the quality control parameters to work with
+# the scale of the raw data.
 hourlymax = np.mean(df.quantile(0.95))
 kt = df / hourlymax
 
+# ################
+# # COMPUTE CMVs #
+# ################
+
+# The cmv.compute_cmv function performs the cross-correlation analysis to
+# determine the best fit wind speed and direction for the data. The method
+# parameter can be used to choose between the two methods.
 cld_spd_gag, cld_dir_gag, dat_gag = cmv.compute_cmv(kt, pos_utm,
                                                     reference_id=None,
                                                     method='gagne')
@@ -18,11 +38,21 @@ cld_spd_jam, cld_dir_jam, dat_jam = cmv.compute_cmv(kt, pos_utm,
                                                     reference_id=None,
                                                     method='jamaly')
 
+# Print out some results for these cases
 print("Method     Speed  Angle:rad  Angle:°   N_good")
 print(f"Gagne   {cld_spd_gag:8.2f} {cld_dir_gag:10.2f} {np.rad2deg(cld_dir_gag):8.2f} {sum(dat_gag.pair_flag == cmv.Flag.GOOD):8,}")
 print(f"Jamaly  {cld_spd_jam:8.2f} {cld_dir_jam:10.2f} {np.rad2deg(cld_dir_jam):8.2f} {sum(dat_jam.pair_flag == cmv.Flag.GOOD):8,}")
 
 
+# ################
+# # PLOT RESULTS #
+# ################
+
+# This plot shows the relationship between the time lag computed from the
+# cross-correlation analysis and the separation distance between each pair in
+# the final CMV-wise direction. Ideally, the time lag should be linearly
+# related to the separation distance, with the slope of the line being the
+# cloud motion speed.
 plt.title('Result Data for Jamaly Method')
 plt.scatter(dat_jam.pair_lag, dat_jam.pair_dists, c=dat_jam.corr_lag, vmin=0, vmax=1, s=1)
 plt.plot([-150, 200], cld_spd_jam * np.array([-150, 200]), 'r--', linewidth=2)  # Best Fit
@@ -33,6 +63,9 @@ plt.xlabel('Time Lag Between Signals (s)')
 plt.ylabel('Windward Separation Distance (m)')
 plt.legend(['Point Correlation', 'Speed Fit'])
 
+# This plot shows the same data as above, but highlights the points that were
+# flagged as GOOD by the quality control routine. These points are used to
+# determine the actual best fit cloud motion speed and direction.
 plt.figure()
 plt.title('Highlighted GOOD Points for Jamaly Method')
 plt.scatter(dat_jam.pair_lag, dat_jam.pair_dists, c=dat_jam.corr_lag, vmin=0, vmax=1, s=1)
@@ -47,7 +80,9 @@ plt.ylabel('Windward Separation Distance (m)')
[…]
