### 2024-07-16 917472e
Hopefully fix failing CI tests. Appears to be related to a numpy version pinning issue from pandas?
