### 2024-07-16 9d6563e
Updates to the README particularly in response to the issues with the mission statement and contributing sections.
Changes to demos reflect information adjusted in the readme describing variable names, etc.
