+++ docs/sphinx/source/conf.py
@@ -25,6 +25,7 @@ extensions = ['sphinx.ext.duration',
               'sphinx.ext.autodoc',
               'sphinx.ext.autosummary',
               'sphinx.ext.intersphinx',
+              'nbsphinx',
               ]
 
 intersphinx_mapping = {
@@ -39,9 +40,12 @@ templates_path = ['_templates']
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = 'alabaster'
+html_theme = 'sphinx_rtd_theme'
 html_static_path = ['_static']
 
+# Control the display of class members
+autodoc_member_order = 'bysource'
 
 # -- Options for EPUB output
 epub_show_urls = 'footnote'
+
