+++ docs/sphinx/source/examples.rst
@@ -1,0 +1,26 @@
+CMV Examples
+------------
+
+.. toctree::
+   :maxdepth: 1
+
+   demos/cmv_demo
+   demos/automate_cmv_demo
+
+Field Analysis Examples
+-----------------------
+
+.. toctree::
+   :maxdepth: 1
+
+   demos/field_demo
+   demos/field_demo_detailed
+   demos/field_reassignment_demo
+
+Other Examples
+--------------
+
+.. toctree::
+   :maxdepth: 1
+
+   demos/plant_digitization_example
