+++ docs/sphinx/source/field.rst
@@ -1,7 +1,5 @@
-solartoolbox.field module
--------------------------
+Plant Field Validation
+----------------------
 
 .. automodule:: solartoolbox.field
-   :members:
-   :undoc-members:
-   :show-inheritance:
+   :members: compute_predicted_position, compute_delays, assign_positions, remap_positions, cascade_remap
