### 2024-07-18 cbbb277
Draft of sphinx docs for the functions
