+++ docs/sphinx/source/conf.py
@@ -41,7 +41,7 @@ templates_path = ['_templates']
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
 html_theme = 'sphinx_rtd_theme'
-html_static_path = ['_static']
+# html_static_path = ['_static']
 
 # Control the display of class members
 autodoc_member_order = 'bysource'
