+++ src/solartoolbox/signalproc.py
@@ -27,17 +27,23 @@ def correlation(baseline, estimation, scaling='coeff'):
     estimation : numeric or pandas.Series
         The predicted signal
 
-    scaling : a string for the type scaling to use
-        Either 'energy', 'coeff', 'unbiased_energy', 'unbiased_coeff'
-        'energy' - scales by the energy of the autocorrelation of the input
-        'coeff' - scales the output to the correlation coefficient (always
-                  removes the mean from both signals)
-        'unbiased_energy' - similar to energy, but normalizes based on lag to
-                            account for the fewer points used in the
-                            convolution. Removes bias towards small lags.
-        'unbiased_coeff' - similar to coeff, but normalizes based on lag to
-                            account for the fewer points used in the
-                            convolution. Removes bias towards small lags.
+    scaling : str
+        a string for the type scaling to use.
+
+        Options are:
+            `energy`:
+                scales by the energy of the autocorrelation of the input
+            `coeff`:
+                scales the output to the correlation coefficient (always
+                removes the mean from both signals)
+            `unbiased_energy`:
+                similar to energy, but normalizes based on lag to account for
+                the fewer points used in the convolution. Removes bias towards
+                small lags.
+            `unbiased_coeff`:
+                similar to coeff, but normalizes based on lag to
+                account for the fewer points used in the
+                convolution. Removes bias towards small lags.
 
     Returns
     -------
@@ -375,8 +381,10 @@ def tf_delay(tf, coh, coh_limit=0.6, freq_limit=0.02, method='fit'):
         The frequency limit to use for filtering the phase. Set to None to
         include all points.
     method : str
-        The method to use for computing the delay. Options are:
-            'diff' - unwrap phase and take derivative (*not well validated)
+        The method to use for computing the delay.
+
+        Options are:
+            'diff' - unwrap phase and take derivative (not well validated)
             'fit' - fit a line to the phase
             'multi' - fit a line to the phase for multiple tfs at once
 
@@ -519,17 +527,22 @@ def xcorr_delay(ts_in, ts_out, scaling='coeff'):
     ts_out : pd.Series
         The output timeseries
     scaling : str
-        Type of scaling to use for cross correlation. Options are:
+        Type of scaling to use for cross correlation.
 
-        'energy' - scales by the energy of the autocorrelation of the input
-        'coeff' - scales the output to the correlation coefficient (always
-                  removes the mean from both signals)
-        'unbiased_energy' - similar to energy, but normalizes based on lag to
-                            account for the fewer points used in the
-                            convolution. Removes bias towards small lags.
-        'unbiased_coeff' - similar to coeff, but normalizes based on lag to
-                            account for the fewer points used in the
-                            convolution. Removes bias towards small lags.
+        Options are:
+            `energy`:
+                scales by the energy of the autocorrelation of the input
+            `coeff`:
+                scales the output to the correlation coefficient (always
+                removes the mean from both signals)
+            `unbiased_energy`:
+                similar to energy, but normalizes based on lag to account for
+                the fewer points used in the convolution. Removes bias towards
+                small lags.
+            `unbiased_coeff`:
+                similar to coeff, but normalizes based on lag to account for
+                the fewer points used in the convolution. Removes bias towards
+                small lags.
 
     Returns
     -------
@@ -558,16 +571,27 @@ def compute_delays(ts_in, ts_out, mode='loop', scaling='coeff'):
     Parameters
     ----------
     ts_in: pd.DataFrame
+        The time series representing the input signal
     ts_out: pd.DataFrame
+        The time series representing the output signal
     mode: str (default 'loop')
[…]
