+++ docs/sphinx/source/generated/solartoolbox.cmv.rst
@@ -1,37 +1,0 @@
-﻿solartoolbox.cmv
-================
-
-.. automodule:: solartoolbox.cmv
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      compute_cmv
-      optimum_subset
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      Flag
-      WindspeedData
-   
-   
-
-   
-   
-   
-
-
-
