+++ docs/sphinx/source/generated/solartoolbox.rst
@@ -1,23 +1,0 @@
-﻿solartoolbox
-============
-
-.. automodule:: solartoolbox
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
