+++ docs/sphinx/source/api.rst
@@ -1,11 +1,0 @@
-Other Functions
-===============
-
-.. autosummary::
-    :toctree: generated
-
-    solartoolbox.cmv
-    solartoolbox.field
-
-.. toctree::
-    solartoolbox
