+++ docs/sphinx/source/generated/solartoolbox.field.rst
@@ -1,33 +1,0 @@
-﻿solartoolbox.field
-==================
-
-.. automodule:: solartoolbox.field
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      assign_positions
-      cascade_remap
-      compute_delays
-      compute_predicted_position
-      remap_positions
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-

[…]
