### 2024-07-18 jranalli/solarspatialtools#7 @jranalli
Thanks for these suggestions, it was motivation for me to learn how to use sphinx, though I admit in a very basic way. I built some API documentation that also gives some organization to the docs that I hope will address some of these comments. Changes are in the referencing pull request.

I'm hesitating a bit to publish these right to a readthedocs project with #8 outstanding. I agree in principle that the name is somewhat generic. I'm not really sure whether there are potential consequences of changing it since it already exists on pypi, and I don't want to create that same problem on readthedocs if the name is up in the air. I was thinking of running the name discussion by the reviewers collectively to get feedback on whether a name change is recommended before trying to resolve that.

Given that: @cwhanse for now do you feel it's reasonable to generate the sphinx docs locally for the referenced PR to take a look and see if they address your concerns? If not I'll gladly just jump and get them on readthedocs ASAP.
