### 2024-07-18 jranalli/solarspatialtools#7 @cwhanse
I think its better to get agreement on the package name, and then I can look at the docs.
