### 2024-07-18 #6984 @jranalli
@cwhanse raised [an issue](https://github.com/jranalli/solartoolbox/issues/8) about the package name being too generic as well. I'd appreciate additional feedback on that one before making that kind of switch, as it might have implications for how this review process works, since it would rename the github paths, and it would affect the pypi deployment. So I'd ask the other reviewers, when doing the review, if you could please chime in if you have thoughts about the name as well I'd appreciate it. Thanks!
