### 2024-07-18 jranalli/solarspatialtools#8 @jranalli
While I agree in principle that the name is somewhat generic, my personal preference would be to leave it, if only for logistical reasons. But I would be happy to consider further feedback and opinions as part of this process.
