### 2024-07-21 #6984 @AdamRJensen
> @cwhanse raised [an issue](https://github.com/jranalli/solartoolbox/issues/8) about the package name being too generic as well. I'd appreciate additional feedback on that one before making that kind of switch, as it might have implications for how this review process works, since it would rename the github paths, and it would affect the pypi deployment. So I'd ask the other reviewers, when doing the review, if you could please chime in if you have thoughts about the name as well I'd appreciate it. Thanks!

@jranalli if you do decide to change the name, then let's wait until all reviews have been completed. With that said, I'm inclined to agree that it's very generic, and if the package will remain specific to spatial analysis then I think it makes sense to change the name.
