### 2024-07-21 jranalli/solarspatialtools#8 @AdamRJensen
@jranalli if you do decide to change the name, then I would suggest waiting until all reviews have been completed (though I see that @cwhanse suggests otherwise, so your choice). With that said, I'm inclined to agree that it's very generic, and if the package will remain specific to spatial analysis then I think it makes sense to change the name.
