### 2024-07-25 1eddb59
Initialize rename of repository to solarspatialtools. Github and pip renaming pending.
