### 2024-07-25 4b1370a
Trying to fix ipython warnings on readthedocs builds
