### 2024-07-25 jranalli/solarspatialtools#8 @jranalli
I'm going to go ahead with this proposal, but as @AdamRJensen suggested, I will either delay the actual renaming of the GitHub repository, or wait until there's a bit more clarity about potential impacts on the JOSS review process. I believe that there may not be big impacts, because after checking on a test repository, it appears that GitHub automatically redirects things, but I'll follow whatever guidance JOSS has on this.

I don't think I can change over my github actions for the pypi publishing until I've renamed things here on GitHub so that will have to wait until the GitHub renaming is resolved.
