### 2024-07-25 jranalli/solarspatialtools#7 @jranalli
A draft of the documentation for a renamed package based on pull request #10 is available on [readthedocs](https://solarspatialtools--10.org.readthedocs.build/en/10/).
