### 2024-07-25 jranalli/solarspatialtools#8 @jranalli
Installation is currently available via testpypi for anyone who wanted to use that pathway.

`pip install -i https://test.pypi.org/simple/ solarspatialtools`
