### 2024-07-29 jranalli/solarspatialtools#8 @kperrynrel
I'm going through my JOSS review right now, and I am also inclined for changing the name here to solarspatialtools, as solartoolbox is generic and is thrown around a lot for different applications when I do a quick Google search on it. Adding my opinion in here as I know you were interested in getting a consensus on this @jranalli
