### 2024-08-04 jranalli/solarspatialtools#8 @kbonney
Sounds like you have made up your mind @jranalli, but my personal 2 cents is that the name `solartoolbox` while generic, did suit the repository since it is intended to be a catchall for various PV functionality that falls outside of the scope of the major packages. Either way works for me, though.
