+++ joss2024/paper.bib
@@ -181,6 +181,7 @@
 	author = {Lave, Matthew and Reno, Matthew J. and Broderick, Robert J.},
 	year = {2015},
 	pages = {327--337},
+	doi = {10.1016/j.solener.2015.05.028},
 }
 
 
