### 2024-08-08 5a0fb66
Fixes to documentation (#9)
- Fix minor bug in jamaly cmv
- Add demo for plant digitization by @williamhobbs 
- introduce sphinx documentation re: JOSS review commentary
