+++ demos/field_demo_full_process.py
@@ -1,223 +1,0 @@
-import itertools
-
-import pandas as pd
-import numpy as np
-import matplotlib.pyplot as plt
-
-from solartoolbox import cmv, spatial, field
-
-# If tqdm is installed, use it to communicate the progress of the loop
-try:
-    from tqdm import tqdm
-except ImportError:
-    def tqdm(x):
-        return x
-
-
-# This demo shows the entire process of using the field module to remap the
-# combiner positions in a plant layout. This is a more advanced example than
-# the field_demo.py, and it demonstrates how to perform averaging over multiple
-# CMV pairs as proposed in the initial paper on the method and to iteratively
-# remap the combiner positions to find the best possible layout.
-
-
-# #############
-# # READ DATA #
-# #############
-
-datafile = "data/sample_plant_1.h5"
-
-dfs = {}
-cmvs = {}
-
-# Load the plant layout
-pos_utm = pd.read_hdf(datafile, mode="r", key="utm").infer_objects()
-cmb_ids = pos_utm.index
-inv_ids = list(set([nm.split('-')[1] for nm in cmb_ids]))
-inv_ids.sort()
-
-# Load all the time windows. If starting from truly raw data (i.e. time series
-# where potential CMV windows are not already known, you could consider
-# following the methodological approach in automate_cmv_demo.py to determine
-# time periods when the CMV is likely to be useful.
-for key in ['a', 'b', 'c', 'd', 'e']:
-    dfs[key] = pd.read_hdf(datafile, mode="r", key=f"data_{key}").infer_objects()
-
-# ################
-# # COMPUTE CMVs #
-# ################
-
-# Calculate CMVs for each time window
-for key in dfs.keys():
-    cld_spd, cld_dir, dat = cmv.compute_cmv(dfs[key], pos_utm, method='jamaly', options={'minvelocity': 1})
-    cmvs[key] = spatial.pol2rect(cld_spd, cld_dir)
-print(pd.DataFrame(cmvs.values(), index=cmvs.keys(), columns=['cld_x', 'cld_y']))
-
-# Determine suitable CMV pairs based on anglular separation
-min_ang_sep = 45
-cmv_pairs = list(itertools.combinations(cmvs.keys(), 2))
-for pair in cmv_pairs.copy():
-    ang_between = np.rad2deg(
-        np.arccos(np.dot(spatial.unit(cmvs[pair[0]]), spatial.unit(cmvs[pair[1]]))))
-    if not (min_ang_sep < ang_between < (180 - min_ang_sep)):
-        cmv_pairs.remove(pair)
-print(cmv_pairs)
-
-# Create the initial mapping. Initially, we assume all plan positions are
-# correct, and we'll remap them as we go. remap will represent the current
-# mapping of combiners to combiners. We'll start with a 1:1 mapping.
-# pos_map will represent the current plant layout used for this iteration.
-remap = [(ref, ref) for ref in cmb_ids]
-pos_map = pos_utm.copy()
-
-# Make lists to hold the values at each iteration, and store current vals
-positions = []  # A list of the plant layouts
-predictions = []  # A list of combiner position predictions
-remaps = []  # A list of the remapping indices
-positions.append(pos_map)
-remaps.append(remap)
-
-# ###########################
-# # ITERATE TIL CONVERGENCE #
-# ###########################
-
-# Loop until the data converges
-working = True
-while working:
-    # In this loop, we'll compute the positions within the whole plant. This
-    # is the level at which we will calculate perform the remappings, based
-    # on the combiner positions for the whole plant. This loop iterates,
-    # updating predictions every time we get a better idea of the true plant
-    # layout.
-
[…]
