+++ demos/signalproc_demo.py
@@ -1,182 +1,0 @@
-import pandas as pd
-import matplotlib.pyplot as plt
-import numpy as np
-import pvlib.location
-
-from solartoolbox import spatial, signalproc, stats
-
-
-"""
-This demo will compute transfer functions for a plant within the HOPE Melpitz
-dataset. The plant consists of a linear subset of points aligned in the N-S
-direction, and during an hour-long window with known cloud motion vector. 
-
-It will compute some variability statistics and plot the different models of 
-the transfer function.
-"""
-
-
-def main():
-    # ##### INPUTS ##### #
-
-    # The file name
-    fn = "data/hope_melpitz_1s.h5"
-
-    # This model has a known cloud speed and direction during this time
-    twin = pd.date_range('2013-09-08 9:15:00', '2013-09-08 10:15:00',
-                         freq='1s', tz='UTC')
-    cloud_speed = 20  # m/s
-    cloud_dir = (0, 1)  # as x-y vector
-
-    # The reference site within the plant, and the plant subset (N-S)
-    ref_id = 40
-    plant_pts = [60, 79, 92, 43, 42, 40, 86, 7, 14, 70]
-    
-    # ##### DATA SETUP ##### #
-
-    # Read Data from HDF file
-    pos = pd.read_hdf(fn, mode="r", key="latlon")
-    ts = pd.read_hdf(fn, mode="r", key="data")
-
-    # Project them to UTM
-    pos_utm = spatial.latlon2utm(pos['lat'], pos['lon'])
-    # Select just the points in the plant
-    pos_sub = pos.loc[plant_pts]
-    pos_sub_utm = pos_utm.loc[plant_pts]
-
-    # Downselect the time series to just the points in the plant, and just
-    # the time window of interest.
-    ts_sub = ts.loc[twin]
-
-    # Select the Input (reference) time series, and compute the Output
-    # (aggregate time series)
-    ts_in = ts_sub[ref_id]
-    ts_sub = ts_sub[plant_pts]
-    ts_agg = ts_sub.sum(axis=1) / len(plant_pts)
-
-    # Calculate the clear sky index for the time series (using PVLIB functions)
-    loc = pvlib.location.Location(np.mean(pos_sub['lat']),
-                                  np.mean(pos_sub['lon']))
-    cs_ghi = loc.get_clearsky(ts_in.index, model='simplified_solis')['ghi']
-    kt_in = pvlib.irradiance.clearsky_index(ts_in, cs_ghi, 2)
-    kt_agg = pvlib.irradiance.clearsky_index(ts_agg, cs_ghi, 2)
-
-    # ##### TRANSFER FUNCTIONS ##### #
-
-    # Apply the filter from the Cloud Advection Model
-    camfilt = signalproc.get_camfilter(pos_sub_utm, cloud_speed,
-                                       cloud_dir, pos_utm.loc[ref_id])
-    ts_cam = signalproc.apply_filter(ts_in, camfilt)
-    kt_cam = pvlib.irradiance.clearsky_index(ts_cam, cs_ghi)
-
-    # Calculate Marcos Model Filtered Signal
-    marcosfilt = signalproc.get_marcosfilter(10, camfilt.index)
-    ts_marcos = signalproc.apply_filter(ts_in, marcosfilt)
-    kt_marcos = pvlib.irradiance.clearsky_index(ts_marcos, cs_ghi)
-
-    # Calculate WVM filtered signal
-    kt_wvm, _, _ = pvlib.scaling.wvm(kt_in, pos_sub_utm, cloud_speed)
[…]
