### 2024-08-08 #6984 @jranalli
@AdamRJensen question regarding the name change. I'm going to go ahead with that and call it `solarspatialtools` as several suggested.

You had tentatively said that I should hold off on renaming the package until after the JOSS review. I did a little testing and GitHub keeps links to old repository names redirected. See for example [https://github.com/jranalli/readthedocs-tutorial](https://github.com/jranalli/readthedocs-tutorial) and [https://github.com/jranalli/readthedocs-tutorial-renametest](https://github.com/jranalli/readthedocs-tutorial-renametest), which go to the same place. It appears to work for command line clones and links to specific files as well. I don't have that good knowledge of what all of JOSS's infrastructure is doing, so I can't say for sure.

Since a lot of the concerns raised dealt with documentation, I'm just trying to balance between creating new docs that all read the old name, or whether it's ok to move ahead with the name change and get a good updated version of the docs to match for the reviewers to sign off on.
