+++ src/solarspatialtools/dataio/nrcan_data.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import os
 
-from solartoolbox.dataio import iotools
+from solarspatialtools.dataio import iotools
 
 """
 Tool for processing NRCAN spatially resolved data as provided on the NRCAN 
