+++ .readthedocs.yml
@@ -14,6 +14,8 @@ python:
    install:
       - method: pip
         path: .
+        extra_requirements:
+          - docs
 
 sphinx:
   configuration: docs/sphinx/source/conf.py
