### 2024-08-08 5cc204b
v 0.4.2 Rename to SolarSpatialTools
