+++ docs/sphinx/source/conf.py
@@ -6,12 +6,12 @@
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'solartoolbox'
+project = 'solarspatialtools'
 copyright = '2024, Joe Ranalli'
 author = 'Joe Ranalli'
 
 release = '0.4'
-version = '0.4.1'
+version = '0.4.2'
 
 import os
 import sys
