### 2024-08-08 jranalli/solarspatialtools#8 @jranalli
Name changed and pushed. Still will need some checking to ensure that publishing to pypi and readthedocs works.
