### 2024-08-08 7e6d999
Add badges and link to readthedocs to readme
