### 2024-08-08 jranalli/solarspatialtools#7 @jranalli
Alright, having gone through with the rename, the latest published version has tried to address many of these comments, and there are [API docs](https://solarspatialtools.readthedocs.io/en/stable/) corresponding to the latest version along with it. If there are still concerns, please let me know so I can address them.
