### 2024-08-08 a7368a1
Add paper ref to README
