### 2024-08-08 #6984 @jranalli
Ok, the rename has gone through. PyPI has the package under the updated name.
