### 2024-08-08 #6984 @jranalli
Made a couple very small changes to the PDF to accommodate the name change and to add one of the suggested DOIs.
