### 2024-08-09 7824978
Add upper bounds to the specified mathematical library versions.
