+++ setup.py
@@ -20,12 +20,12 @@ setup(
     packages=find_packages(where="src"),
     install_requires=[
         'numpy<2',
-        'pandas',
-        'tables',  # possibly must be installed separately via conda pytables because of MS dependency
-        'pyproj',
-        'pvlib',
-        'netcdf4',
-        'scipy'
+        'pandas<=2.2.2',
+        'tables<=3.9.2',
+        'pyproj<=3.6.1',
+        'pvlib<=0.11.0',
+        'netcdf4<1.7.2',
+        'scipy<=1.14.0'
     ],
     license='BSD (3 Clause)',
     extras_require=dict(tests=['pytest'], demos=['matplotlib', 'jupyter'], docs=['sphinx', 'nbsphinx', 'sphinx_rtd_theme', 'ipython']),
