### 2024-08-09 jranalli/solarspatialtools#7 @cwhanse
@jranalli the revisions answer my concerns about documentation.
