### 2024-08-19 def9f90
Modify github actions to run all tests in one action.
