### 2024-08-19 654b810
Add matplotlib dependency and extra demo install instructions to readme
