+++ setup.py
@@ -25,8 +25,9 @@ setup(
         'pyproj<=3.6.1',
         'pvlib<=0.11.0',
         'netcdf4<1.7.2',
-        'scipy<=1.14.0'
+        'scipy<=1.14.0',
+        'matplotlib<=3.9.2',
     ],
     license='BSD (3 Clause)',
-    extras_require=dict(tests=['pytest'], demos=['matplotlib', 'jupyter'], docs=['sphinx', 'nbsphinx', 'sphinx_rtd_theme', 'ipython']),
+    extras_require=dict(tests=['pytest'], demos=['jupyter'], docs=['sphinx', 'nbsphinx', 'sphinx_rtd_theme', 'ipython']),
 )
