### 2024-08-19 d6b480c
Catching one extra instance of solarspatialtools name
