### 2024-08-19 10423a6
Move demo files to utilize nblink with sphinx
