+++ docs/sphinx/source/demos/field_reassignment_demo.py
@@ -1,165 +1,0 @@
-import pandas as pd
-import numpy as np
-import matplotlib.pyplot as plt
-
-from solarspatialtools import spatial, field
-
-# #############
-# # READ DATA #
-# #############
-
-# This example demonstrates how the automated remapping can be used to
-# determine the best possible positions for each combiner in a plant layout.
-# We'll skip over some of the details of explaining the field calculation,
-# please refer to the field_demo.py for a more detailed explanation.
-
-# Load the data for Sample Plant 1 and specify the CMVs
-datafile = "data/sample_plant_1.h5"
-ts_data_a = pd.read_hdf(datafile, mode="r", key="data_a").infer_objects()
-ts_data_c = pd.read_hdf(datafile, mode="r", key="data_c").infer_objects()
-cmv_a = spatial.pol2rect(10.7, 3.29)
-cmv_c = spatial.pol2rect(3.14, 1.92)
-
-# Load the plant layout
-pos_utm = pd.read_hdf(datafile, mode="r", key="utm").infer_objects()
-
-# We now have two hours of data, each with a different CMV. These two are known
-# to have roughly perpendicular directions so the field analysis should work.
-
-# Inverter 20 is known to have an incorrect layout, so we'll focus on those
-# combiners for this demonstration
-refs = [nm for nm in pos_utm.index if nm.split('-')[1] in ['20']]
-
-# #############################
-# # PERFORM FIELD CALCULATION #
-# #############################
-
-# The first step is to perform the field calculation for the combiners that
-# we're interested in. This will give us the expected position of the combiners
-# relative to these particular CMVs.
-
-# Perform the calculation for each combiner and store the results
-preds_r0 = pd.DataFrame(index=refs, columns=['E', 'N'], dtype=float)
-for ref in refs:
-    pos, _ = field.compute_predicted_position([ts_data_a, ts_data_c], pos_utm, ref, [cmv_a, cmv_c])
-    preds_r0.loc[ref] = pos
-
-# field.assign_positions will return a list of tuples that can be used to
-# remap the combiner positions to the new, best-fit positions.
-# The tuple is written as (Combiner X, Copy-Location-From)
-remap_inds_r0, _ = field.assign_positions(pos_utm.loc[refs], preds_r0)
-print(str(remap_inds_r0) + '\n')
-
-# ############################
-# # REPEAT FIELD CALCULATION #
-# ############################
-
-# Because the combiner position predictions depend on their original positions,
-# we need to re-run the remapping process, with these remapped positions in
-# place, to see to what extent the current prediction was dependent upon those
-# positions. So we'll run as follows.
-
-# Create a remapped version of the plant layout using the remapped indices
-pos_utm_r1 = field.remap_positions(pos_utm, remap_inds_r0)
-
-# Rerun the field calculation with the remapped positions
-preds_r1 = pd.DataFrame(index=refs, columns=['E', 'N'], dtype=float)
-for ref in refs:
-    pos, _ = field.compute_predicted_position([ts_data_a, ts_data_c], pos_utm_r1, ref, [cmv_a, cmv_c])
-    preds_r1.loc[ref] = pos
-
-# Update the assignment of combiner positions based on this recalculation.
-remap_inds_r1, _ = field.assign_positions(pos_utm_r1.loc[refs], preds_r1)
-
[…]
