+++ docs/sphinx/source/demos/field_demo_full_process_multithread.py
@@ -1,257 +1,0 @@
-import itertools
-from multiprocessing import Pool
-
-import pandas as pd
-import numpy as np
-import matplotlib.pyplot as plt
-
-from solarspatialtools import cmv, spatial, field
-
-# This demonstration is identical to field_demo_full_process.py with the
-# exception that it uses the multiprocessing module to parallelize the
-# calculation of the combiner positions. This can be useful for large plants
-# where the calculation of the combiner positions is the bottleneck.
-# Testing on my laptop showed a roughly 4x speedup for the Sample Plant 1 data
-# (2:30 per iteration serially vs about 0:40 in the multithreaded case).
-
-
-def main():
-    """
-    Function to wrap the calculations. We can't do multithreaded processing
-    without putting it inside a function like this and using:
-      if __name__ == "__main__":
-        main()
-    """
-    # #############
-    # # READ DATA #
-    # #############
-
-    datafile = "data/sample_plant_1.h5"
-
-    ts_datas = {}
-    cmvs = {}
-
-    # Load the plant layout
-    pos_utm = pd.read_hdf(datafile, mode="r", key="utm").infer_objects()
-    cmb_ids = pos_utm.index
-    inv_ids = list(set([nm.split('-')[1] for nm in cmb_ids]))
-    inv_ids.sort()
-
-    # Load all the time windows. If starting from truly raw data (i.e. time series
-    # where potential CMV windows are not already known, you could consider
-    # following the methodological approach in automate_cmv_demo.py to determine
-    # time periods when the CMV is likely to be useful.
-    for key in ['a', 'b', 'c', 'd', 'e']:
-        ts_datas[key] = pd.read_hdf(datafile, mode="r", key=f"data_{key}").infer_objects()
-
-    # ################
-    # # COMPUTE CMVs #
-    # ################
-
-    # Calculate CMVs for each time window
-    for key in ts_datas.keys():
-        cld_spd, cld_dir, dat = cmv.compute_cmv(ts_datas[key], pos_utm, method='jamaly', options={'minvelocity': 1})
-        cmvs[key] = spatial.pol2rect(cld_spd, cld_dir)
-    print(pd.DataFrame(cmvs.values(), index=cmvs.keys(), columns=['cld_x', 'cld_y']))
-
-    # Determine suitable CMV pairs based on anglular separation
-    min_ang_sep = 45
-    cmv_pairs = list(itertools.combinations(cmvs.keys(), 2))
-    for pair in cmv_pairs.copy():
-        ang_between = np.rad2deg(
-            np.arccos(np.dot(spatial.unit(cmvs[pair[0]]), spatial.unit(cmvs[pair[1]]))))
-        if not (min_ang_sep < ang_between < (180 - min_ang_sep)):
-            cmv_pairs.remove(pair)
-    print(cmv_pairs)
-
-    # Create the initial mapping. Initially, we assume all plan positions are
-    # correct, and we'll remap them as we go. remap will represent the current
-    # mapping of combiners to combiners. We'll start with a 1:1 mapping.
-    # pos_map will represent the current plant layout used for this iteration.
-    remap = [(ref, ref) for ref in cmb_ids]
-    pos_map = pos_utm.copy()
-
-    # Make lists to hold the values at each iteration, and store current vals
-    positions = []  # A list of the plant layouts
-    predictions = []  # A list of combiner position predictions
-    remaps = []  # A list of the remapping indices
-    positions.append(pos_map)
-    remaps.append(remap)
-
-    # ###########################
-    # # ITERATE TIL CONVERGENCE #
-    # ###########################
-
-    # Loop until the data converges
-    working = True
-    while working:
-        # In this loop, we'll compute the positions within the whole plant. This
-        # is the level at which we will calculate perform the remappings, based
[…]
