+++ docs/sphinx/source/conf.py
@@ -26,6 +26,7 @@ extensions = ['sphinx.ext.duration',
               'sphinx.ext.autosummary',
               'sphinx.ext.intersphinx',
               'nbsphinx',
+              'nbsphinx_link',
               ]
 
 intersphinx_mapping = {

[…]
