### 2024-08-19 b7602b5
adjust tests to point to moved demo file locations
