+++ joss2024/paper.md
@@ -26,22 +26,24 @@ bibliography: paper.bib
 
 # Summary
 
-Solar energy is a form of renewable energy whose resource (i.e. sunlight) is available for capture on the earth's surface with a fixed energy density. This type of resource inherently requires spatial distribution of collection infrastructure in order to achieve increased generation scale. This is true both in the case of distributed (e.g. rooftop solar) and centralized generation. As international responses to climate change promote growing interest in solar energy, there is a corresponding growth of interest in tools for working with distributed solar energy data that possesses these characteristics. This package, `SolarSpatialTools` aims to contribute to that need by providing research codes for spatial analyses of solar energy data and resources. 
+Solar energy is a form of renewable energy whose resource (i.e. sunlight) is available for capture on the earth's surface with a fixed energy density. This type of resource inherently requires spatial distribution of collection infrastructure in order to achieve increased generation scale. This is true both in the case of distributed (e.g. rooftop solar) and centralized generation. As international responses to climate change promote growing interest in solar energy, there is a corresponding growth of interest in tools for working with distributed solar energy data that possesses these characteristics. This package, `SolarSpatialTools`, aims to contribute to that need by providing research codes for spatial analyses of solar energy data and resources. 
 
 # Statement of need
 
-As mature packages already exist for supporting general analysis and modeling of solar energy systems, such as `pvlib-python` [@anderson_pvlib_2023] and `pvanalytics` [@perry_pvanalytics_2022], this package is not intended to serve as a replacement, a competitor or to fragment those communities. Rather, `SolarSpatialTools` serves to collect codes for several tasks that are out-of-scope for `pvlib-python` and `pvanalytics`, but are still of general interest to the research community. Where appropriate, capabilities of `SolarSpatialTools` are contributed to `pvlib-python` or `pvanalytics`. For example, a python language port of the Wavelet Variability Model [@lave_cloud_2013] contained in the MATLAB `pvlib` package [@andrews_introduction_2014] was first developed within `SolarSpatialTools` but was contributed to `pvlib-python` in 2019. `SolarSpatialTools` primarily grew out of personal research codes developed by the lead author under the name `solartoolbox`, but as tools have reached a level of maturity that attracted interest of a broader audience, it has been prepared as a package for more general public use.
+As mature packages already exist for supporting general analysis and modeling of solar energy systems, such as `pvlib-python` [@anderson_pvlib_2023] and `pvanalytics` [@perry_pvanalytics_2022], this package is not intended to serve as a replacement, a competitor, or to fragment those communities. Rather, `SolarSpatialTools` serves to collect codes for several tasks that are out-of-scope for `pvlib-python` and `pvanalytics`, but are still of general interest to the research community. Where appropriate, capabilities of `SolarSpatialTools` are contributed to `pvlib-python` or `pvanalytics`. For example, a python language port of the Wavelet Variability Model [@lave_cloud_2013] contained in the MATLAB `pvlib` package [@andrews_introduction_2014] was first developed within `SolarSpatialTools` but was contributed to `pvlib-python` in 2019. `SolarSpatialTools` primarily grew out of personal research codes developed by the lead author under the name `solartoolbox`, but as tools have reached a level of maturity that attracted interest of a broader audience, it has been prepared as a package for more general public use.
 
[…]
