### 2024-09-13 e0a4bd3
Update paper.md
