+++ joss2024/paper.md
@@ -32,7 +32,7 @@ Solar energy is a form of renewable energy whose resource (i.e. sunlight) is ava
 
 As mature packages already exist for supporting general analysis and modeling of solar energy systems, such as `pvlib-python` [@anderson_pvlib_2023] and `pvanalytics` [@perry_pvanalytics_2022], this package is not intended to serve as a replacement, a competitor, or to fragment those communities. Rather, `SolarSpatialTools` serves to collect codes for several tasks that are out-of-scope for `pvlib-python` and `pvanalytics`, but are still of general interest to the research community. Where appropriate, capabilities of `SolarSpatialTools` are contributed to `pvlib-python` or `pvanalytics`. For example, a python language port of the Wavelet Variability Model [@lave_cloud_2013] contained in the MATLAB `pvlib` package [@andrews_introduction_2014] was first developed within `SolarSpatialTools` but was contributed to `pvlib-python` in 2019. `SolarSpatialTools` primarily grew out of personal research codes developed by the lead author under the name `solartoolbox`, but as tools have reached a level of maturity that attracted interest of a broader audience, it has been prepared as a package for more general public use.
 
-To be more specific, a variety of analytical techniques related to solar energy are documented in literature, but are not already implemented by existing packages in part due to their relatively high complexity relative to the those packages' intended scope. For example, techniques for processing cloud motion vectors (CMVs) from spatially distributed data sets are documented in the literature, such as the method by Jamaly and Kleissl [@jamaly_robust_2018] and that by Gagne et al. [@gagne_directional_2018]. Implementation of these techniques is laborious, requiring calculation of mutual correlation between all possible sensor pairs within a distributed data set. This fundamentally leads to a need to handle data types (i.e. simultaneous time series for each sensor) that are not aligned with the primary focus of the existing packages. Further, the number of calculation steps that are specialized for these CMV calculations makes them unattractive for inclusion in existing solar energy packages, without leading to an extreme broadening of scope to adapt to this singular use case. At the same time, the level of detail in those calculation steps makes them potentially difficult for other investigators to individually implement on a consistent and optimized basis. As they serve a common need within solar energy research, they are implemented in a well documented way by `SolarSpatialTools` to help alleviate this challenge.
+To be more specific, a variety of analytical techniques related to solar energy are documented in literature, but are not already implemented by existing packages in part due to their relatively high complexity relative to those packages' intended scope. For example, techniques for processing cloud motion vectors (CMVs) from spatially distributed data sets are documented in the literature, such as the method by Jamaly and Kleissl [@jamaly_robust_2018] and that by Gagne et al. [@gagne_directional_2018]. Implementation of these techniques is laborious, requiring calculation of mutual correlation between all possible sensor pairs within a distributed data set. This fundamentally leads to a need to handle data types (i.e. simultaneous time series for each sensor) that are not aligned with the primary focus of the existing packages. Further, the number of calculation steps that are specialized for these CMV calculations makes them unattractive for inclusion in existing solar energy packages, without leading to an extreme broadening of scope to adapt to this singular use case. At the same time, the level of detail in those calculation steps makes them potentially difficult for other investigators to individually implement on a consistent and optimized basis. As they serve a common need within solar energy research, they are implemented in a well documented way by `SolarSpatialTools` to help alleviate this challenge.
 
 # Features
 
@@ -50,13 +50,13 @@ These three main capabilities are also supported by extended documentation and t
[…]
