### 2024-09-13 #6984 @AdamRJensen
@jranalli could you upload a version of the latest release of the package to Zenodo and post the doi here?
