### 2024-09-13 #6984 @cwhanse
> @cwhanse is there a doi for this report: https://energy.sandia.gov/wp-content/gallery/uploads/Stein_ASES_2012_VI_paper_SAND2012-2088C.pdf?

There isn't. @jsstein we should get one. This conference paper is posted on a Sandia page, OSTI and in ResearchGate.
