### 2024-09-13 3922dab
Minor paper revisions (#17)
Paper revisions proposed by editor.
