### 2024-09-13 9079ef0
Minor rephrase for clarity. Revision of module vs package terminology
