+++ joss2024/paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 # Summary
 
-Solar energy is a form of renewable energy whose resource (i.e. sunlight) is available for capture on the earth's surface with a fixed energy density. This type of resource inherently requires spatial distribution of collection infrastructure in order to achieve increased generation scale. This is true both in the case of distributed (e.g. rooftop solar) and centralized generation. As international responses to climate change promote growing interest in solar energy, there is a corresponding growth of interest in tools for working with distributed solar energy data that possesses these characteristics. This package, `SolarSpatialTools`, aims to contribute to that need by providing research codes for spatial analyses of solar energy data and resources. 
+Solar energy is a form of renewable energy whose resource (i.e. sunlight) is available on the earth's surface with a relatively low energy density. This type of resource inherently requires spatial distribution of collection infrastructure in order to achieve increased generation scale. This is true both in the case of distributed (e.g. rooftop solar) and centralized generation. As international responses to climate change promote growing interest in solar energy, there is a corresponding growth of interest in tools for working with distributed solar energy data that possesses these characteristics. This package, `SolarSpatialTools`, aims to contribute to that need by providing research codes for spatial analyses of solar energy data and resources. 
 
 # Statement of need
 
@@ -36,36 +36,36 @@ To be more specific, a variety of analytical techniques related to solar energy 
 
 # Features
 
-There are three capabilities of the `SolarSpatialTools` package that are most likely to be of interest for a general audience. These main capabilities are contained in the following packages:
+There are three capabilities of the `SolarSpatialTools` package that are most likely to be of interest for a general audience. These main capabilities are contained in the following modules:
 
 - `signalproc`: tools for performing signal processing analyses across multi-sensor networks of solar energy data
 - `cmv`: tools for computing the cloud motion vector from spatially distributed sensor networks 
 - `field`: tools for analyzing the relative positions of spatially distributed measurement units via cloud motion
 
-These three main capabilities are also supported by extended documentation and tutorials in an additional package:
+These three main capabilities are also supported by extended documentation and tutorials in an additional directory of the package:
 
 - `demos`: demonstration codes and sample data to help users get started with the package
 
 ## Signal Processing
-The `signalproc` package was developed as part of efforts to analyze aggregation of irradiance by spatially distributed plants, but may also be applicable to other signal processing tasks. This approach is used by the Wavelet Variability Model [@lave_cloud_2013], the model of Marcos et al. [@marcos_power_2011] and the Cloud Advection Model [@ranalli_cloud_2021], which was developed by the lead author based on the physical intuition of Hoff and Perez [@hoff_quantifying_2010]. The package contains codes for implementing these types of models using a transfer function paradigm. Some wrappers are provided for `scipy` [@virtanen_scipy_2020] signal processing functions to simplify their application on the data type conventions used by this package. A demonstration of the signal processing capability as it pertains to comparing the different spatial aggregation models is provided in the `demos` directory of the package (`signalproc_demo.py`).
[…]
