### 2024-09-13 af8f6bd
Add dates to changelog, closes #18
