+++ Changelog.md
@@ -12,35 +12,35 @@ the primary impact being on my individual use. Beginning with version 0.4.0,
 the package will adopt a consistent approach and will strive to maintain 
 backwards compatibility for releases within the same minor version (0.X.X). 
 
-# Version 0.2
+# Version 0.2 - Jun 15, 2022
 First public release of the package via PyPi
-# Version 0.2.1
+# Version 0.2.1 - Jul 17, 2022
 - Add wrapper for `pvlib.clearsky_index` to handle pandas type
-# Version 0.2.2
+# Version 0.2.2 - July 27, 2022
 - Change input to camfilter to handle references that don't coincide with the 
 site itself. 
 - ** Breaks backwards compatibility **
 # Version 0.2.3
 - Add methods for calculating delay between signals to `signalproc`
-# Version 0.2.4
+# Version 0.2.4 - Oct 19, 2023
 - Add some additional options to `cmv` for qc and filtering of the data
 # Version 0.2.5
 - Add location analysis in `field` package
-# Version 0.3.1
+# Version 0.3.1 - Nov 8, 2023
 - Expands `field` analysis capability
 - Add comprehensive testing
 - ** Breaks backwards compatibility **
-# Version 0.3.2
+# Version 0.3.2 - Nov 14, 2023
 - Major speed improvements via vectorization of CMV 
 - ** Breaks backwards compatibility due to function signature changes **
-# Version 0.3.3
+# Version 0.3.3 - Nov 17, 2023
 - Major speed improvements to field analysis (vectorization of transfer function delays)
 - ** Breaks backwards compatibility **
-# Version 0.3.4
+# Version 0.3.4 - Nov 21, 2023
 - Bug fix for handling and excluding NaN within `field`
-# Version 0.3.5
+# Version 0.3.5 - Dec 18, 2023
 - Add multi-column support to several functions in `stats`. Thanks to Scott Sheppard for the suggestion.
-# Version 0.4.0
+# Version 0.4.0 - Apr 10, 2024
 - Added some functions to `field` to support automated remapping.
     - See `field_reassignment_demo` for details.
 - Added a function to `cmv` for identifying an optimum subset of CMV pairs. 
@@ -54,16 +54,16 @@ site itself.
     - `field_demo_full_process` - Overview of the full field processing workflow 
     - `field_demo_full_process_multithread` - Same as previous but including multithreading support for speed
 - Added tests of these new functions for more completeness.
-# Version 0.4.1
+# Version 0.4.1 - Aug 8, 2024
 - Bug fix for `cmv` so that `jamaly` r_corr is always positive. 
 - Demo for digitizing a plant equipment layout added by @williamhobbs
 - Improve comments in `cmv_demo`
 - Add sphinx documentation for entire library
-# Version 0.4.2
+# Version 0.4.2 - Aug 8 - 2024
 - Rename the library to solarspatialtools
-# Version 0.4.3 & 0.4.4
+# Version 0.4.3 & 0.4.4 - Aug 8, 2024
 - Fixes to GitHub deployment issues, no functional changes
-# Version 0.4.5
+# Version 0.4.5 - Sept 13, 2024
 - Improvements to documentation
 - Simplification of github actions
 - Added matplotlib dependency
