### 2024-09-13 7227153
Switchover to setup.cfg only. Regarding #20.
