### 2024-09-13 7f1f965
Documentation Improvements regarding JOSS Reviews (#13)
* Improvements to documentation
* Pinned dependencies to avoid test errors
* Add matplotlib dependency
* Move demos back to parent directory
* Switch to setup.cfg for package description
* Changes for JOSS review

---------

Co-authored-by: Joe Ranalli <jranalli@psu.edu>
