+++ docs/sphinx/source/conf.py
@@ -11,7 +11,7 @@ copyright = '2024, Joe Ranalli'
 author = 'Joe Ranalli'
 
 release = '0.4'
-version = '0.4.4'
+version = '0.4.5'
 
 import os
 import sys
@@ -26,6 +26,7 @@ extensions = ['sphinx.ext.duration',
               'sphinx.ext.autosummary',
               'sphinx.ext.intersphinx',
               'nbsphinx',
+              'nbsphinx_link',
               ]
 
 intersphinx_mapping = {
@@ -48,4 +49,7 @@ autodoc_member_order = 'bysource'
 
 # -- Options for EPUB output
 epub_show_urls = 'footnote'
+
+# Hide warning per sphinx #12300, appears to come from nbsphinx or nbsphinx_link
+suppress_warnings = ["config.cache"]
 
