+++ tests/test_cmv.py
@@ -97,7 +97,7 @@ def test_cmv_artificial(theta_deg, velocity, mode):
 
 
 def test_cmv_gagne_data():
-    datafile = "../docs/sphinx/source/demos/data/sample_plant_1.h5"
+    datafile = "../demos/data/sample_plant_1.h5"
     pos_utm = pd.read_hdf(datafile, mode="r", key="utm")
     df = pd.read_hdf(datafile, mode="r", key="data_a")
 
@@ -115,7 +115,7 @@ def test_cmv_gagne_data():
 
 
 def test_cmv_jamaly_data():
-    datafile = "../docs/sphinx/source/demos/data/sample_plant_1.h5"
+    datafile = "../demos/data/sample_plant_1.h5"
     pos_utm = pd.read_hdf(datafile, mode="r", key="utm")
     df = pd.read_hdf(datafile, mode="r", key="data_a")
 
@@ -133,7 +133,7 @@ def test_cmv_jamaly_data():
 def test_cmv_jamaly_data_ref():
     # A test using a reference when we calculate the CMV
 
-    datafile = "../docs/sphinx/source/demos/data/sample_plant_1.h5"
+    datafile = "../demos/data/sample_plant_1.h5"
     pos_utm = pd.read_hdf(datafile, mode="r", key="utm")
     df = pd.read_hdf(datafile, mode="r", key="data_a")
 
