+++ setup.cfg
@@ -1,2 +1,43 @@
 [metadata]
-description_file = README.md
+name = solarspatialtools
+version = 0.4.5
+author = 'Joe Ranalli'
+author_email = jar339@psu.edu
+description = A package for spatial analysis of solar energy and solar irradiance
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/jranalli/solarspatialtools
+project_urls =
+    Bug Tracker = https://github.com/jranalli/solarspatialtools/issues
+    Documentation = https://solarspatialtools.readthedocs.io
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: BSD (3 Clause)
+    Operating System :: OS Independent
+    Topic :: Scientific/Engineering
+    Intended Audience :: Science/Research
+
+[options]
+install_requires =
+    numpy<2
+    pandas<=2.2.2
+    tables<=3.9.2
+    pyproj<=3.6.1
+    pvlib<=0.11.0
+    netcdf4<1.7.2
+    scipy<=1.14.0
+    matplotlib<=3.9.2
+
+[options.extras_require]
+tests =
+    pytest
+
+demos =
+    jupyter
+
+docs =
+    sphinx
+    sphinx_rtd_theme
+    nbsphinx
+    ipython
+    nbsphinx-link

[…]
