### 2024-09-13 806d047
Fix license label issue in cfg.
