+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = solarspatialtools
-version = 0.4.5
+version = 0.4.6
 author = 'Joe Ranalli'
 author_email = jar339@psu.edu
 description = A package for spatial analysis of solar energy and solar irradiance
@@ -12,7 +12,7 @@ project_urls =
     Documentation = https://solarspatialtools.readthedocs.io
 classifiers =
     Programming Language :: Python :: 3
-    License :: OSI Approved :: BSD (3 Clause)
+    License :: OSI Approved :: BSD License
     Operating System :: OS Independent
     Topic :: Scientific/Engineering
     Intended Audience :: Science/Research
