+++ setup.cfg
@@ -36,8 +36,8 @@ demos =
     jupyter
 
 docs =
-    sphinx
-    sphinx_rtd_theme
-    nbsphinx
-    ipython
-    nbsphinx-link
+    sphinx==7.4.7
+    sphinx_rtd_theme==2.0.0
+    nbsphinx==0.9.5
+    ipython==8.27.0
+    nbsphinx-link==1.3.0
