+++ setup.cfg
@@ -19,14 +19,14 @@ classifiers =
 
 [options]
 install_requires =
-    numpy<2
-    pandas<=2.2.2
-    tables<=3.9.2
-    pyproj<=3.6.1
-    pvlib<=0.11.0
-    netcdf4<1.7.2
-    scipy<=1.14.0
-    matplotlib<=3.9.2
+    numpy
+    pandas
+    tables
+    pyproj
+    pvlib
+    netcdf4
+    scipy
+    matplotlib
 
 [options.extras_require]
 tests =
