### 2024-09-14 f3e9049
Try tests on multiple python versions and os's
