+++ .github/workflows/tests.yml
@@ -5,7 +5,14 @@ on: [pull_request, push]
 jobs:
   tests:
 
-    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false  # don't cancel other matrix jobs when one fails
+      matrix:
+        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+        os: [ubuntu-latest, macos-latest, windows-latest]
+
+
+    runs-on: ${{ matrix.os }}
 
     steps:
     - name: Checkout
@@ -14,7 +21,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v5
       with:
-        python-version: "3.x"
+        python-version: ${{ matrix.python-version }}
 
     - name: Install dependencies
       run: |
