### 2024-09-14 7ca8fcc
Cleaning Up Dependency Pins and Test Targets
- Updating dependencies  and test targets
- Fix deprecation warnings raised by numpy
