+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = solarspatialtools
-version = 0.4.6
+version = 0.4.7
 author = 'Joe Ranalli'
 author_email = jar339@psu.edu
 description = A package for spatial analysis of solar energy and solar irradiance
@@ -19,14 +19,14 @@ classifiers =
 
 [options]
 install_requires =
-    numpy<2
-    pandas<=2.2.2
-    tables<=3.9.2
-    pyproj<=3.6.1
-    pvlib<=0.11.0
-    netcdf4<1.7.2
-    scipy<=1.14.0
-    matplotlib<=3.9.2
+    numpy
+    pandas
+    tables
+    pyproj
+    pvlib
+    netcdf4
+    scipy
+    matplotlib
 
 [options.extras_require]
 tests =
@@ -36,8 +36,8 @@ demos =
     jupyter
 
 docs =
-    sphinx
-    sphinx_rtd_theme
-    nbsphinx
-    ipython
-    nbsphinx-link
+    sphinx==7.4.7
+    sphinx_rtd_theme==2.0.0
+    nbsphinx==0.9.5
+    ipython==8.27.0
+    nbsphinx-link==1.3.0
