### 2024-09-14 #6984 @jranalli
> @jranalli could you upload a version of the latest release of the package to Zenodo and post the doi here?

@AdamRJensen are you satisfied with all the issues you had raised? If so, I'll get right on this, so that the one archived coincides with what passed review.
