### 2024-09-15 c831786
Finalized JOSS publication version
