+++ Changelog.md
@@ -70,6 +70,7 @@ site itself.
 - Move demos back to the parent directory using nblink
 # Version 0.4.6 - Sept 13, 2024
 - Fix license name in cfg
-# Version 0.4.7 - Sept 14, 2024
+# Version 0.5.0 - Sept 15, 2024
 - Updating dependencies  and test targets
 - Fix deprecation warnings raised by numpy
+- JOSS publication
