+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = solarspatialtools
-version = 0.4.7
+version = 0.5.0
 author = 'Joe Ranalli'
 author_email = jar339@psu.edu
 description = A package for spatial analysis of solar energy and solar irradiance
