### 2024-09-16 #6984 @kyleniemeyer
Hi @jranalli, can you add your coauthor to the software archive on Zenodo? You can update the metadata there without having to release a new version.

In the paper:
- Please add commas after instances of "e.g." and "i.e", and make sure Python is capitalized throughout.
- Where you have author names listed immediately prior to the citation (e.g., "Jamaly and Kleissl (Jamaly & Kleissl, 2018)"), please use this citation format (and remove the manual addition of names): `@author:2001`  ->  "Author et al. (2001)"
- When you have multiple citations, please use this format to combine them: `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
